sam0/adc: merged samd21 & saml21 adc driver
