+++ boards/saml21-xpro/include/periph_conf.h
@@ -122,18 +122,13 @@ static const spi_conf_t spi_config[] = {
  * @name ADC Configuration
  * @{
  */
-#define ADC_NUMOF                          ADC_0_CHANNELS
-#define ADC_0_EN                           1
-#define ADC_MAX_CHANNELS                   14
-/* ADC 0 device configuration */
-#define ADC_0_DEV                          ADC
-#define ADC_0_IRQ                          ADC_IRQn
-#define ADC_0_CHANNELS                     (3U)
+#define ADC_NUMOF                          (3U)
 
 /* ADC 0 Default values */
 #define ADC_0_CLK_SOURCE                   0 /* GCLK_GENERATOR_0 */
 #define ADC_0_PRESCALER                    ADC_CTRLB_PRESCALER_DIV256
-static const adc_channel_t adc_channels[] = {
+
+static const adc_conf_chan_t adc_channels[] = {
     /* port, pin, muxpos */
     {GPIO_PIN(PA, 10), ADC_INPUTCTRL_MUXPOS(ADC_INPUTCTRL_MUXPOS_AIN18)},
     {GPIO_PIN(PA, 11), ADC_INPUTCTRL_MUXPOS(ADC_INPUTCTRL_MUXPOS_AIN19)},
